Main

Documentation

outStr :: [String] -> String

main :: IO ()

tarefa :: [String] -> [String]

type Entrada = [String]

type Saida = String

type Tabuleiro = [String]

type Coordenadas = String

type Comandos = String

verificaTab :: Entrada -> Saida

verificaTab recebe o input. Caso o tabuleiro não seja válido dá como resultado a linha em que ocorre o erro. Se for válido (ou seja, se for igual a 0 (zero)) verifica o que está abaixo do tabuleiro.

Exemplo:

>>> verificaTab ["aacdD","aacaa","bbCaa","0 1 S","SEASLEAADSAL"]
1. validaTab ["aacdD","aacaa","bbCaa"]
2. verificaCoord ["0 1 S","SEASLEAADSAL"] 5 3 (3+1)

validaTab :: Tabuleiro -> Int

validaTab verifica se todas as linhas do tabuleiro tem o mesmo tamanho. Se sim devolve o valor 0 (zero). Se não, devolve o número da linha em que ocorre o erro.

Exemplo:

>>> validaTab ["aacdD","aacaa","bbCaa"] = 0

verificaCoord :: [String] -> Int -> Int -> Int -> Saida

verificaCoord verifica se as coordenadas são válidas. Se forem verifica o que está abaixo. Se não forem devolve o número da linha a que corresponde a linha das coordenadas.

Exemplo:

>>> verificaCord ["0 1 S","SEASLEAADSAL"] 5 3 4
1. validaCoord 5 3 "0 1 S"
2. verificaCom ["SEASLEAADSAL"] (4+1)

validaCoord :: Int -> Int -> Coordenadas -> Bool

validaCoord verifica se as coordenadas são maiores ou iguais que 0 e menores que o tamanho do tabuleiro e verifica se a orientação é válida.

Exemplo:

>>> validaCoord 5 3 "0 1 S" = True

verificaCom :: [String] -> Int -> Saida

verificaCom verifica se os comandos são válidos. Caso não sejam devolve o número da linha a que corresponde a linha dos comandos. Se forem verifica ainda se o que está abaixo é vazio'. Se não houver nada abaixo dos comandos significa que o tabuleiro é válido. Se houver devolve o número a que corresponde a primeira linha abaixo dos comandos.

Exemplo:

>>> verificaCom ["SEASLEAADSAL"] 5
1. validaComand "SEASLEAADSAL"

validaComand :: Comandos -> Bool

validaComand verifica se a linha dos comando é válida.

Exemplo:

>>> validaComand "SEASLEAADSAL" = True