Main

Documentation

type Tabuleiro = [String]

type Coordenada = Int

type Coordenadas = (Coordenada, Coordenada)

type Orientacao = Char

type Comandos = String

t1 :: [[Char]]

t2 :: [[Char]]

t3 :: [[Char]]

t4 :: [[Char]]

t5 :: [[Char]]

t6 :: [[Char]]

t7 :: [[Char]]

t8 :: [[Char]]

t9 :: [[Char]]

t10 :: [[Char]]

t11 :: [[Char]]

t12 :: [[Char]]

t13 :: [[Char]]

t14 :: [[Char]]

t15 :: [[Char]]

t16 :: [[Char]]

t17 :: [[Char]]

t18 :: [[Char]]

t19 :: [[Char]]

outStr :: [String] -> String

main :: IO ()

tarefa :: [String] -> [String]

prefixo :: [[Char]]

criaTab :: Tabuleiro -> [String]

criaTab, aux, linhas, mostra e mostraP calculam as coordenadas de todas as posições do tabuleiro.

Exemplo

>>> criaTab ["aacdD","aacaa","bbCaa"]
1. aux ["aacdD","aacaa","bbCaa"] 0 2

aux :: Tabuleiro -> Coordenada -> Coordenada -> [String]

Exemplo

'>>>aux' [aacdD,aacaa,bbCaa] 0 2 1. linhas aacdD 0 2 ++ aux [aacaa,bbCaa] 0 1

linhas :: String -> Coordenada -> Coordenada -> [String]

Exemplo

>>> linhas "aacdD" 0 2
1. mostra 0 2 'a' ++ linhas "acdD" 1 2

mostra :: Coordenada -> Coordenada -> Char -> [String]

Exemplo

>>> mostra 0 2 'a'
1. nivel 'a' == 0 = [mostraP 0 2 'a'] 
2. [mostraP 0 2 'a'] ++ mostra x y (chr (ord (toLower 'a') - 1))

mostraP :: Coordenada -> Coordenada -> Char -> String

Exemplo

>>> mostraP 0 2 'a' = "<Transform translation='" ++ (show ((fromIntegral 0)*3.5) ++ " " ++ show ((fromIntegral 2)*3.5) ++
                     " " ++ show (nivel 'a')) ++ "'> <Shape USE=\"tile\"/> </Transform>"

nivel :: Char -> Float

nivel calcula a altura de um caracter do tabuleiro.

Exemplo

>>> nivel 'a' = (fromIntegral (ord (toLower 'a') - ord 'a'))*1.5

robot :: Tabuleiro -> Coordenadas -> Orientacao -> [String]

robot calcula as coordenadas e a orientação iniciais do robot.

Exemplo

>>> robot ["aacdD","aacaa","bbCaa"] (0,1) 'S' = ["<Transform translation='" ++ (show ((fromIntegral 0)*3.5) ++ " " ++
                                                show ((fromIntegral 1)*3.5) ++ " " ++ 
                                                show (altura ["aacdD","aacaa","bbCaa"] (0,1))) ++ 
                                                "' rotation='0 0 1 " ++ show (orientacao 'S') ++ "'>",
                                                "<Transform translation='0 0 2.2' rotation='1 0 0 1.57'>"]

altura :: Tabuleiro -> Coordenadas -> Float

altura calcula a altura de uma determinada coordenada.

Exemplo

>>> altura ["aacdD","aacaa","bbCaa"] (0,1)
1. 1.5 + nivel 'a'

orientacao :: Orientacao -> Float

prefixoanimacao :: [[Char]]

animacao :: Tabuleiro -> Coordenadas -> Orientacao -> Comandos -> [String]

key :: (Int, Float) -> Float -> String

key calcula a fracao de tempo que dura cada comando.

Exemplo

>>> key (12,1/12) 1
1. " " ++ show (1*(1/12)) ++ key (11,1/12) 2

extra :: String -> Int

extra calcula o numero de comandos Luz e Saltar que existem num comando.

Exemplo

>>> extra "SEASLEAADSAL" = length "LL" + length "SSS"

proxsPos :: Tabuleiro -> Coordenadas -> Orientacao -> Comandos -> (String, String)

proxsPos calcula o movimento de cada comando.

Exemplo

>>> proxsPos ["aacdD","aacaa","bbCaa"] (0,1) 'S' "SEASLEAADSAL"
1. ((proxPosSul ["aacdD","aacaa","bbCaa"] (0,1) 'S' "SEASLEAADSAL" [",","0","0","0"]),(roda "SEASLEAADSAL" [",","0","0","0","0"])

proxPosSul :: Tabuleiro -> Coordenadas -> Orientacao -> Comandos -> [String] -> String

proxPosSul calcula os movimentos para os comandos Saltar, Avançar e Luz.

Exemplo

>>> proxPosSul ["aacdD","aacaa","bbCaa"] (0,1) 'S' "SEASLEAADSAL" [",","0","0","0"]
1. presaltoS ++ unwoords salto ++ proxPosSul ["aacdD","aacaa","bbCaa"] (proxPosAS) 'S' "SEASLEAADSAL" salto

proxPosSA :: Coordenadas -> Orientacao -> (Int, Int)

proxPosSA calcula as próximas coordenadas para os comandos Saltar e Avançar.

Exemplo

>>> proxPosSA (0,1) 'S' = (0,0)

validaCoord :: (Int, Int) -> Coordenadas -> Bool

validaCoord verifica se a próxima coordendada calculada é válida.

Exemplo

>>> validaCoord (5,3) (0,0) = True

validaS :: Char -> Char -> Bool

validaS verifica se é possível saltar desde a coordenada inicial para a próxima coordenada calculada, comparando os caracteres correspondentes.

Exemplo

>>> validaS 'A' 'B' = True

sobe :: [Float] -> [String]

sobe acrescenta 1.5 à altura.

Exemplo

>>> sobe [0,0,0] = [",","0","1.5","0"]

desceS :: [Float] -> Orientacao -> [String]

desceS calcula o movimento necessário para avançar para de seguida descer.

Exemplo

>>> desceS [0,0,0] 'S' = [",","0","0","3.5"]

proxPosSS :: [Float] -> Char -> Char -> Orientacao -> [String]

proxPosSS calcula o movimento para o comando saltar.

Exemplo

>>> proxPosSS [0,0,0] 'a' 'b' 'S' = [",","0","1.5","3.5"]

proxPosA :: [Float] -> Orientacao -> [String]

proxPosA calcula o movimento para o comando Avançar.

Exemplo

>>> proxPosA [0,0,0] 'S' = [",","0","0","3.5"]

proxPosL :: [Float] -> String

proxPosL calcula o movimento para o comando Luz (dar um pulo)

Exemplo

>>> proxPosL [0,0,0] = unwords [",","0","1.5","0"] ++ unwords [",","0","0","0"]

proxPosE :: Char -> Char

proxPosE e proxPosD calculam a orientação para os comandos esquerda e direita, respetivamente.

Exemplo

>>> proxPosE 'S' = 'E'

proxPosD :: Char -> Char

Exemplo

>>> proxPosD 'S' = 'O'

roda :: Comandos -> [String] -> String

roda calcula todos os movimentos de rotações para os comandos esquerda e direita.

Exemplo

>>> roda "SEASLEAADSAL" [",","0","0","0","0"]
1. unwords [",","0","0","0","0"] ++ roda "EASLEAADSAL" [",","0","0","0","0"]

proxPosEste :: Tabuleiro -> Coordenadas -> Orientacao -> Comandos -> [String] -> String

proxPosEste calcula os movimentos para os comandos Saltar, Avançar e Luz.

Exemplo

>>> proxPosEste ["aacdD","aacaa","bbCaa"] (0,1) 'E' "SEASLEAADSAL" [",","0","0","0"]
1. presaltoS ++ unwoords salto ++ proxPosEste ["aacdD","aacaa","bbCaa"] (proxPosAS) 'E' "SEASLEAADSAL" salto

proxPosSE :: [Float] -> Char -> Char -> Orientacao -> [String]

proxPosSE calcula o movimento para o comando saltar.

Exemplo

>>> proxPosSE [0,0,0] 'a' 'b' 'E' = [",","0","1.5","3.5"]

proxPosAE :: [Float] -> Orientacao -> [String]

proxPosAE calcula o movimento para o comando Avançar.

Exemplo

>>> proxPosAE [0,0,0] 'E' = [",","0","0","3.5"]

desceE :: [Float] -> Orientacao -> [String]

desceE calcula o movimento necessário para avançar para de seguida descer.

Exemplo

>>> desceE [0,0,0] 'E' = [",","0","0","3.5"]

proxPosNorte :: Tabuleiro -> Coordenadas -> Orientacao -> Comandos -> [String] -> String

proxPosNorte calcula os movimentos para os comandos Saltar, Avançar e Luz.

Exemplo

>>> proxPosNorte ["aacdD","aacaa","bbCaa"] (0,1) 'N' "SEASLEAADSAL" [",","0","0","0"]
1. presaltoS ++ unwoords salto ++ proxPosNorte ["aacdD","aacaa","bbCaa"] (proxPosAS) 'N' "SEASLEAADSAL" salto

proxPosSN :: [Float] -> Char -> Char -> Orientacao -> [String]

proxPosSN calcula o movimento para o comando saltar.

Exemplo

>>> proxPosSN [0,0,0] 'a' 'b' 'N' = [",","0","1.5","3.5"]

proxPosAN :: [Float] -> Orientacao -> [String]

proxPosAN calcula o movimento para o comando Avançar.

Exemplo

>>> proxPosAN [0,0,0] 'N' = [",","0","0","3.5"]

desceN :: [Float] -> Orientacao -> [String]

desceN calcula o movimento necessário para avançar para de seguida descer.

Exemplo

>>> desceN [0,0,0] 'N' = [",","0","0","3.5"]

proxPosOeste :: Tabuleiro -> Coordenadas -> Orientacao -> Comandos -> [String] -> String

proxPosOeste calcula os movimentos para os comandos Saltar, Avançar e Luz.

Exemplo

>>> proxPosOeste ["aacdD","aacaa","bbCaa"] (0,1) 'O' "SEASLEAADSAL" [",","0","0","0"]
1. presaltoS ++ unwoords salto ++ proxPosOeste ["aacdD","aacaa","bbCaa"] (proxPosAS) 'O' "SEASLEAADSAL" salto

proxPosSO :: [Float] -> Char -> Char -> Orientacao -> [String]

proxPosSO calcula o movimento para o comando saltar.

Exemplo

>>> proxPosSO [0,0,0] 'a' 'b' 'O' = [",","0","1.5","3.5"]

proxPosAO :: [Float] -> Orientacao -> [String]

proxPosAO calcula o movimento para o comando Avançar.

Exemplo

>>> proxPosAO [0,0,0] 'O' = [",","0","0","3.5"]

desceO :: [Float] -> Orientacao -> [String]

desceO calcula o movimento necessário para avançar para de seguida descer.

Exemplo

>>> desceO [0,0,0] 'O' = [",","0","0","3.5"]

sufixoanimacao :: [[Char]]

sufixoRo :: [[Char]]

sufixo :: [[Char]]