Main

Documentation

outStr :: [String] -> String

main :: IO ()

tarefa :: [String] -> [String]

type Tabuleiro = [String]

type Comandos = String

type Coordenadas = [String]

type Luzes = [String]

Lista de coordenadas onde o comando L é executado com sucesso, mas que não tem coordenadas repetidas

type LuzesTotal = [String]

Lista de coordenadas onde o comando L é executado com sucesso

type Contagem = Int

type Resultado = [String]

type Comando = Char

type Acumulador = (Coordenadas, Luzes, LuzesTotal, Contagem)

type CoordenadaX = Int

type CoordenadaY = Int

type Orientacao = Char

type CaracterTab = Char

type LuzesTabuleiro = [String]

Lista de coordenadas em que existe lâmpada no tabuleiro

type Dimensao = (Int, Int)

Dimensão do tabuleiro

proxsPos :: Tabuleiro -> LuzesTabuleiro -> Dimensao -> Comandos -> Acumulador -> Resultado

proxsPos calcula a proxima posicao usando a posicao calculada com o comando anterior (proxPos). No entanto antes de calcular a próxima posição verifica se todas as luzes estão acesas. Caso estejam termina e mostra o resultado.

Exemplos:

>>> proxsPos ["aacdD","aacaa","bbCaa"] ["0 2","4 2"] (5,3) "SEASLEAADSAL" (["0","1","S"],[],[],0)
1. pertence ["0 2","4 2"] [] = False
2. proxsPos ["aacdD","aacaa","bbCaa"] ["0 2","4 2"] (5,3) "EASLEAADSAL" (proxPos ["aacdD","aacaa","bbCaa"] (5,3) 'S' (["0","1","S"],[],[],0))

>>> proxsPos ["aacdD","aacaa","bbCaa"] ["0 2","4 2"] (5,3) "EAADSAL" (["2","0","E"],["2 0"],["2 0"],5)
1. pertence ["0 2","4 2"] ["2 0"] = False
2. proxsPos ["aacdD","aacaa","bbCaa"] ["0 2","4 2"] (5,3) "AADSAL" (proxPos ["aacdD","aacaa","bbCaa"] (5,3) 'E' (["2","0","E"],["2 0"],["2 0"],5))

proxPos :: Tabuleiro -> Dimensao -> Comando -> Acumulador -> Acumulador

proxPos calcula a próxima posição dependendo do comando que seja para executar. Caso o comando seja aplicável, mostra a nova posição, e acresenta 1 (um) ao contador. No caso de o comando ser L (e for válido), verifica se a coordenada em que está já foi executada (para efeitos de comparação com as coordenadas existentes no tabuleiro) e também acrescenta essa cooordenada à lista de coordenadas que já passaram por este processo. Caso o comando não há alterações.

Exemplos:

>>> proxPos ["aacdD","aacaa","bbCaa"] (5,3) 'S' (["0","1","S"],[],[],0)
1. verificaS = validaCoord 5 3 proxPosAS
2. (["0","0","S"],[],[],0+1)

>>> proxPos ["aacdD","aacaa","bbCaa"] (5,3) 'L' (["2","0","E"],[],[],4)
1. isUpper 'C' = True
2. (["2","0","E"],verificaRep "0 2" [],[]++"2 0",4+1)

proxPosSA :: CoordenadaX -> CoordenadaY -> Orientacao -> Coordenadas

proxPosSA calcula a próxima posição para o comando saltar e avançar dependendo da sua orientação

Exemplo:

>>> proxPosSA 0 1 S = "0 0 S"

validaCoord :: Int -> Int -> Coordenadas -> Bool

validaCoord verifica se a uma coordenada é válida, isto é, se o x e o y são maiores ou iguais a zero e menores que a dimensão do tabuleiro.

Exemplo:

>>> validaCoordf 5 3 ["0","0","S"] = True

validaS :: CaracterTab -> CaracterTab -> Bool

validaS verifica se o caracter final é o seguinte do inicial ou se é menor.

Exemplo:

>>> validaS 'a' 'b' = True
1. ord 'b' = ord 'a' + 1

proxPosD :: CoordenadaX -> CoordenadaY -> Orientacao -> Coordenadas

proxPosD calcula a proxima posição para o comando direita.

Exemplo:

>>> proxPosD 0 1 S = "0 1 O"

proxPosE :: CoordenadaX -> CoordenadaY -> Orientacao -> Coordenadas

proxPosE calcula a próxima posição para o comando esquerda.

Exemplo:

>>> proxPosE 0 1 S = "0 1 E"

pertence :: Luzes -> Luzes -> Bool

pertence recebe a lista de coordenadas de luzes que existem no tabuleiro e verifica se todas as coordenadas dessa lista pertencem à lista das coordenadas que só tem as luzes que estão acesas.

Exemplo:

>>> pertence ["2 0","4 2"] ["2 0","4 2"] = True

>>> pertence ["2 0","4 2"] ["2 0"] = False

luzesTab :: Tabuleiro -> Luzes

luzesTab,coordLuzes e linhas calcula a lista de coordenadas onde existe letra maiúscula no tabuleiro, isto é onde é preciso acender luz.

Exemplo:

>>> luzesTab ["aacdD","aacaa","bbCaa"]
1. coordLuzes ["aacdD","aacaa","bbCaa"] 2

coordLuzes :: Tabuleiro -> Int -> Luzes

Exemplo:

>>> coordLuzes ["aacdD","aacaa","bbCaa"] 2
1. linhas "aacdD" 0 2 ++ coordLuzes ["aacaa","bbCaa"] (2-1)

linhas :: String -> Int -> Int -> Luzes

Exemplo:

>>> linhas "aacdD" 0 2 = ["4 2"]

verificaRep :: String -> Luzes -> Luzes

verificaRep recebe a coordenada em que foi executado o comando L com sucesso e verifica se essa coordenada já existe na lista das coordenadas que só tem as luzes que estão acesas.

Exemplo:

>>> verificaRep "4 2" ["2 0"] = ["2 0","4 2"]

>>> verificaRep "2 0" ["2 0] = []