Main

Documentation

t1 :: [[Char]]

t2 :: [[Char]]

t3 :: [[Char]]

t4 :: [[Char]]

t5 :: [[Char]]

t6 :: [[Char]]

t7 :: [[Char]]

outStr :: [String] -> String

main :: IO ()

tarefa :: [String] -> [String]

type Coordenadas = (Int, Int)

type Tabuleiro = [String]

type Comando = String

type Orientacao = Char

luzesTab :: Tabuleiro -> [Coordenadas]

luzesTab,coordLuzes e linhas calcula a lista de coordenadas onde existe letra maiúscula no tabuleiro, isto é, onde é preciso acender luz.

Exemplo:

>>> luzesTab ["aaaaA","cAcaa","acaAa","aaaaa"]
1. coordLuzes ["aaaaA","cAcaa","acaAa","aaaaa"] 3

coordLuzes :: Tabuleiro -> Int -> [Coordenadas]

Exemplo:

>>> coordLuzes ["aaaaA","cAcaa","acaAa","aaaaa"] 2
1. linhas "aaaaA" 0 3 ++ coordLuzes ["cAcaa","acaAa","aaaaa"] (3-1)

linhas :: String -> Int -> Int -> [Coordenadas]

Exemplo:

>>> linhas "aaaaA" 0 3 = [(4,3)]

fun :: [Coordenadas] -> [[Coordenadas]] -> [[Coordenadas]]

fun,auxP,aux e aux2 calculam todas as combinações possíveis entre as coordenadas das lâmpadas.

Exemplo

>>> fun[(4,3),(1,2),(3,1)] []
1. fun [(1,2),(3,1)] (auxP (4,3) [])

auxP :: Coordenadas -> [[Coordenadas]] -> [[Coordenadas]]

Exemplo

>>> auxP (4,3) []
1. [[(4,3)]]

>>> auxP (1,2) [[(4,3)]]
1. aux (1,2) [(4,3)] 0

aux :: Coordenadas -> [Coordenadas] -> Int -> [[Coordenadas]]

Exemplo

>>> aux (1,2) [(4,3)] 0
1. [aux2 (1,2) [(4,3)] 0] ++ aux (1,2) [(4,3)] (0+1)

aux2 :: Coordenadas -> [Coordenadas] -> Int -> [Coordenadas]

Exemplo

>>> aux2 (1,2) [(4,3)] 0 = [(1,2),(4,3)]

>>> aux2 (1,2) [(4,3)] 1
1. (4,3): (aux2 (1,2) [] (1-1))

menor :: [Comando] -> Comando

menor calcula a String de menor comprimento de uma dada lista de String's

Exemplo

>>> menor ["EAAAEALAEAALEEAAAEAL","EAAAEALDAEAALEAAAEAL","EAAAAEAAALEAEAALEEAEAAL","EAAAAEAAALEAAAEALEAADAL"] =
"EAAAEALAEAALEEAAAEAL"

movePm :: Tabuleiro -> Coordenadas -> Orientacao -> [[Coordenadas]] -> [Comando]

movePm calcula todas as maneiras de acender todas as lâmpadas.

Exemplo

>>> movePm ["aaaaA","cAaaa","acaAa","aaaaa"] (0,0) 'S'
[[(3,1),(1,2),(4,3)],[(1,2),(3,1),(4,3)],[(1,2),(4,3),(3,1)],[(3,1),(4,3),(1,2)],[(4,3),(3,1),(1,2)],[(4,3),(1,2),(3,1)]]
1. move ["aaaaA","cAaaa","acaAa","aaaaa"] (0,0) 'S' [(3,1),(1,2),(4,3)]

movePM :: Tabuleiro -> Coordenadas -> Orientacao -> [[Coordenadas]] -> Comando

movePM calcula a primeira maneira válida de acender todas as lâmpadas.

Exemplo

>>> movePM ["BaBaB","BaBaB","BaBaB","BaBaB","BaBaB"] (4,3) 'S'
[[(0,4),(2,4),(4,4),(0,3),(2,3),(4,3),(0,2),(2,2),(4,2),(0,1),(2,1),(4,1),(0,0),(2,0),(4,0)],...]
1. move ["BaBaB","BaBaB","BaBaB","BaBaB","BaBaB"] (4,3) 'S' 
[(0,4),(2,4),(4,4),(0,3),(2,3),(4,3),(0,2),(2,2),(4,2),(0,1),(2,1),(4,1),(0,0),(2,0),(4,0)]

move :: Tabuleiro -> Coordenadas -> Orientacao -> [Coordenadas] -> Comando

move calcula os comandos necessários para acender todas as luzes. Se não for possível devolve 1.

Exemplo

>>> move ["aaaaA","cAaaa","acaAa","aaaaa"] (0,0) 'S' [(3,1),(1,2),(4,3)]
1. orientX ["aaaaA","cAaaa","acaAa","aaaaa"] (0,0) 'S' (3,1) ++ orientY ["aaaaA","cAaaa","acaAa","aaaaa"] (3,0) 'E' (3,1)
 ++ "L"
2. orientY ["aaaaA","cAaaa","acaAa","aaaaa"] (0,0) 'S' (3,1) ++ orientX ["aaaaA","cAaaa","acaAa","aaaaa"] (0,1) 'N' (3,1)
 ++ "L"

orientX :: Tabuleiro -> Coordenadas -> Orientacao -> Coordenadas -> Comando

orientX calcula os comandos necessários para mover o robot até igualar a coordenada x da lâmpada que pretende atingir.

Exemplo

>>> orientX ["aaaaA","cAaaa","acaAa","aaaaa"] (0,0) 'S' (3,1)
1. rodaE 'S' ++ moveASE ["aaaaA","cAaaa","acaAa","aaaaa"] (0,0) (3,1)

rodaE :: Orientacao -> Comando

rodaE calcula os comandos que são necessários executar para que o robot fique virado para Este.

Exemplo

>>> rodaE 'S' = "E"

rodaO :: Orientacao -> Comando

rodaO calcula os comandos que são necessários executar para que o robot fique virado para Oeste.

Exemplo

>>> rodaO 'S' = "D"

moveASE :: Tabuleiro -> Coordenadas -> Coordenadas -> Comando

moveASE calcula os comandos necessários, um a um, para mover o robot para Este até igualar a coordenada x da lâmpada que pretende atingir.

Exemplo

>>> moveASE ["aaaaA","cAaaa","acaAa","aaaaa"] (0,0) (3,1)
1. verificaAS ["aaaaA","cAaaa","acaAa","aaaaa"] (0,0) (1,0) ++ moveASE ["aaaaA","cAaaa","acaAa","aaaaa"] (1,0) (3,1)

moveASO :: Tabuleiro -> Coordenadas -> Coordenadas -> Comando

moveASO calcula os comandos necessários, um a um, para mover o robot para Oeste até igualar a coordenada x da lâmpada que pretende atingir.

Exemplo

>>> moveASO ["aaaaA","cAaaa","acaAa","aaaaa"] (4,0) (3,1)
1. verificaAS ["aaaaA","cAaaa","acaAa","aaaaa"] (4,0) (3,0) ++ moveASO ["aaaaA","cAaaa","acaAa","aaaaa"] (3,0) (3,1)

verificaAS :: Tabuleiro -> Coordenadas -> Coordenadas -> Comando

verificaAS calcula se para ir de uma coordenada para uma adjacente é neccesário Avançar ou Saltar, ou então se não é possível.

Exemplo

>>> 'verificaAS' ["aaaaA","cAaaa","acaAa","aaaaa"] (4,0) (3,0)
1. caracter1 == 'A' && caracter2 == 'A' = "A"
2. caracter1 == 'B' && caracter2 == 'A' = "S"
3. caracter1 == 'A' && carcater2 == 'C' = "1"

orientY :: Tabuleiro -> Coordenadas -> Orientacao -> Coordenadas -> Comando

orientY calcula os comandos necessários para mover o robot até igualar a coordenada y da lâmpada que pretende atingir.

Exemplo

>>> orientY ["aaaaA","cAaaa","acaAa","aaaaa"] (3,0) 'E' (3,1)
1. rodaN 'E' ++ moveASN ["aaaaA","cAaaa","acaAa","aaaaa"] (3,0) (3,1)

rodaN :: Orientacao -> Comando

rodaN calcula os comandos que são necessários executar para que o robot fique virado para Norte.

Exemplo

>>> rodaN 'S' = "EE"

rodaS :: Orientacao -> Comando

rodaS calcula os comandos que são necessários executar para que o robot fique virado para Sul.

Exemplo

>>> rodaS 'S' = ""

moveASN :: Tabuleiro -> Coordenadas -> Coordenadas -> Comando

moveASN calcula os comandos necessários, um a um, para mover o robot para Norte até igualar a coordenada y da lâmpada que pretende atingir.

Exemplo

>>> moveASN ["aaaaA","cAaaa","acaAa","aaaaa"] (3,0) (3,1)
1. verificaAS ["aaaaA","cAaaa","acaAa","aaaaa"] (3,0) (3,1) ++ moveASN ["aaaaA","cAaaa","acaAa","aaaaa"] (3,1) (3,1)

moveASS :: Tabuleiro -> Coordenadas -> Coordenadas -> Comando

moveASS calcula os comandos necessários, um a um, para mover o robot para Sul até igualar a coordenada y da lâmpada que pretende atingir.

Exemplo

>>> moveASS ["aaaaA","cAaaa","acaAa","aaaaa"] (3,2) (3,1)
1. verificaAS ["aaaaA","cAaaa","acaAa","aaaaa"] (3,2) (3,1) ++ moveASS ["aaaaA","cAaaa","acaAa","aaaaa"] (3,1) (3,1)